requirements.txt
vk_api
python-dotenv
peewee
requests